mantenimiento de permisos en base a los roles y al menu

+++ application/views/admin/clientes/list.php
@@ -15,7 +15,9 @@
             <div class="box-body">
                 <div class="row">
                     <div class="col-md-12">
+                        <?php if($permisos->insert == 1):?>
                         <a href="<?php echo base_url();?>mantenimiento/clientes/add" class="btn btn-primary btn-flat"><span class="fa fa-plus"></span>Agregar Clietes</a>
+                        <?php endif;?>
                     </div>
                     <hr>
                     <div class="row">
@@ -51,8 +53,12 @@
                                                         <button type="button" class="btn btn-info btn-view-cliente" data-toggle='modal' data-target='#modal-default' value="<?php echo $cliente->cliente;?>">
                                                                 <span class="fa fa-eye"></span>
                                                         </button>
+                                                        <?php if($permisos->update == 1):?>
                                                             <a href="<?php echo base_url();?>mantenimiento/clientes/edit/<?php echo $cliente->cliente?>" class="btn btn-warning"><span class="fa fa-pencil"></span></a>
+                                                        <?php endif;?>
+                                                        <?php if($permisos->delete == 1):?>
                                                             <a href="<?php echo base_url();?>mantenimiento/clientes/delete/<?php echo $cliente->cliente?>" class="btn btn-danger btn-remove"><span class="fa fa-trash"></span></a>
+                                                        <?php endif;?>
                                                     </div>
                                                 </td>
                                             </tr>
